[Test (integration) + Implement] 'infer' PORO collection
Part of #11

'respond_with' for a successful 'GET' request of a collection.
The collection here must be a Ruby object that acts as a collection of DummyPoro and that responds to enumerable methods.
This is because a bare Array wouldn't respond to model_name which is necessary here for inferring the representer class.
When the representer_class is specified and not inferred, an Array still
will do.